major refactoring and Bugfixes
