+++ altlas/Assets/scripts/MapData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 580c1c17cfa3d44909ca79b5a7285484
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
