+++ altlas/Assets/scripts/MapScript.cs
@@ -1,0 +1,7 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MapScript : MonoBehaviour {
+    public MapData data;
+}
